requirements.txt
django>=4.2.0
django-rest-framework>=3.14.0
openai>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
google-auth-oauthlib>=1.0.0
google-auth>=2.22.0
google-api-python-client>=2.100.0
social-auth-app-django>=5.3.0